requirements.txt
flask 
flask_cors
ibmcloudant>=0.5.0
python-dotenv
bcrypt
transformers
Pillow
torch
gunicorn
seaborn 
pandas
numpy
matplotlib
#dotenv

ibm_cloud_sdk_core
ibm_watson
wordcloud

ibm-cos-sdk
